refactor: add bottom component
