+++ src/components/ConstructorGame/ConstructorGame.module.css
@@ -67,15 +67,6 @@ div .word {
   border: none;
 }
 
-/* .word__wrapper {
-  display: flex;
-  list-style-type: none;
-  justify-content: center;
-
-  margin: 1rem 0;
-  width: 100%;
-} */
-
 .word__container {
   display: flex;
   justify-content: center;
@@ -88,25 +79,6 @@ div .word {
   margin-left: 130px;
 }
 
-/* .word__char {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  margin: 0 1rem;
-  height: 50px;
-  width: 60px;
-
-  background: #733999;
-  border-radius: 10px;
-
-  color: #ffffff;
-  font-size: 2.4rem;
-  line-height: 2.8rem;
-  text-decoration: none;
-  cursor: pointer;
-} */
-
 .word__transcription {
   margin: 5px 0;
   font-size: 1.4rem;
@@ -114,24 +86,6 @@ div .word {
   text-align: center;
 
   color: rgba(115, 57, 153, 0.8);
-}
-
-.btn-next {
-  margin: 25px 0 35px;
-  width: 200px;
-  height: 50px;
-
-  background: #ebf3ff;
-  box-shadow: inset -8px -8px 16px #ffffff, inset 8px 8px 16px #f7d2fb;
-  border-radius: 30px;
-  border: none;
-
-  font-weight: bold;
-  font-size: 2.4rem;
-  line-height: 2.8rem;
-  color: #733999;
-
-  filter: drop-shadow(2px 2px 4px rgba(0, 0, 0, 0.25));
 }
 
 .picture {
